task 34 refactoring  front end
