+++ Main/Frontend/voicetotext/src/assets/arrow.png
