+++ Main/Frontend/voicetotext/src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
+import './styles/index.css';
 import '../node_modules/bootstrap/dist/css/bootstrap.min.css';
 import WebFont from 'webfontloader';
