+++ Main/Frontend/voicetotext/src/styles/App.css
@@ -1,6 +1,5 @@
 .Firstpage {
 	height: 102vh;
-	
 }
 
 .Navbarsize {
@@ -16,7 +15,7 @@
 .App-bg {
   margin: 0;
   position: absolute;
-  background-image: url(medical-wallpaper.jpg);
+  background-image: url(./../assets/medical-wallpaper.jpg);
   opacity:0.95;
   width: 100%;  
   height:400px;
