+++ Main/Frontend/voicetotext/src/styles/index.css
